backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
google-adk
pytest
pytest-asyncio
httpx
